refactor: separate data protocols into folders
