+++ src/data/protocols/criptography/Encrypter.d.ts
