+++ src/data/protocols/db/AddAccountRepository.d.ts
@@ -1,0 +1,6 @@
+import { AddAccountModel } from '../../../domain/usecases/AddAccount'
+import AccountModel from '../../../domain/models/Account'
+
+export default interface AddAccountRepository {
+  add: (account: AddAccountModel) => Promise<AccountModel>
+}
