+++ src/data/protocols/db/ErrorLogRepository.d.ts
