+++ src/data/usecases/authenticate-user/DbAuthenticateUser.ts
@@ -1,6 +1,6 @@
 import AuthenticateUser from '../../../domain/usecases/AuthenticateUser'
 import AuthenticateCredentials from '../../../domain/usecases/AuthenticateCredentials'
-import GetAccountByEmailRepository from '../../protocols/GetAccountByEmailRepository'
+import GetAccountByEmailRepository from '../../protocols/db/GetAccountByEmailRepository'
 
 export default class DbAuthenticateUser implements AuthenticateUser {
   constructor (private readonly getAccountByEmailRepository: GetAccountByEmailRepository) {}
