+++ src/infra/criptography/BcryptAdapter.ts
@@ -1,4 +1,4 @@
-import Encrypter from '../../data/protocols/Encrypter'
+import Encrypter from '../../data/protocols/criptography/Encrypter'
 import bcrypt from 'bcrypt'
 
 export default class BcryptAdapter implements Encrypter {
