+++ src/infra/db/mongodb/account-repository/AccountMongoRepository.ts
@@ -1,4 +1,4 @@
-import AddAccountRepository from '../../../../data/protocols/AddAccountRepository'
+import AddAccountRepository from '../../../../data/protocols/db/AddAccountRepository'
 import { AddAccountModel } from '../../../../domain/usecases/AddAccount'
 import AccountModel from '../../../../domain/models/Account'
 import MongoHelper from '../helpers/MongoHelper'
